+++ httpx_auth/oauth2_tokens.py
@@ -4,13 +4,14 @@ import os
 import datetime
 import threading
 import logging
+from typing import Union
 
 from httpx_auth.errors import InvalidToken, TokenExpiryNotProvided, AuthenticationFailed
 
 logger = logging.getLogger(__name__)
 
 
-def decode_base64(base64_encoded_string: str) -> str:
+def _decode_base64(base64_encoded_string: str) -> str:
     """
     Decode base64, padding being optional.
 
@@ -23,8 +24,18 @@ def decode_base64(base64_encoded_string: str) -> str:
     return base64.b64decode(base64_encoded_string).decode("unicode_escape")
 
 
-def is_expired(expiry: float) -> bool:
-    return datetime.datetime.utcfromtimestamp(expiry) < datetime.datetime.utcnow()
+def _is_expired(expiry: float, early_expiry: float) -> bool:
+    return (
+        datetime.datetime.utcfromtimestamp(expiry - early_expiry)
+        < datetime.datetime.utcnow()
+    )
+
+
+def _to_expiry(expires_in: Union[int, str]) -> float:
+    expiry = datetime.datetime.utcnow().replace(
+        tzinfo=datetime.timezone.utc
+    ) + datetime.timedelta(seconds=int(expires_in))
+    return expiry.timestamp()
 
 
 class TokenMemoryCache:
@@ -37,7 +48,7 @@ class TokenMemoryCache:
         self.forbid_concurrent_cache_access = threading.Lock()
         self.forbid_concurrent_missing_token_function_call = threading.Lock()
 
-    def add_bearer_token(self, key: str, token: str):
+    def _add_bearer_token(self, key: str, token: str):
         """
         Set the bearer token and save it
         :param key: key identifier of the token
@@ -49,14 +60,14 @@ class TokenMemoryCache:
             raise InvalidToken(token)
 
         header, body, other = token.split(".")
-        body = json.loads(decode_base64(body))
+        body = json.loads(_decode_base64(body))
         expiry = body.get("exp")
         if not expiry:
             raise TokenExpiryNotProvided(expiry)
 
         self._add_token(key, token, expiry)
 
-    def add_access_token(self, key: str, token: str, expires_in: int):
+    def _add_access_token(self, key: str, token: str, expires_in: int):
         """
         Set the bearer token and save it
         :param key: key identifier of the token
@@ -64,10 +75,7 @@ class TokenMemoryCache:
         :param expires_in: Number of seconds before token expiry
         :raise InvalidToken: In case token is invalid.
         """
-        expiry = datetime.datetime.utcnow().replace(
-            tzinfo=datetime.timezone.utc
-        ) + datetime.timedelta(seconds=int(expires_in))
-        self._add_token(key, token, expiry.timestamp())
+        self._add_token(key, token, _to_expiry(expires_in))
 
     def _add_token(self, key: str, token: str, expiry: float):
         """
@@ -83,12 +91,24 @@ class TokenMemoryCache:
                 f'Inserting token expiring on {datetime.datetime.utcfromtimestamp(expiry)} (UTC) with "{key}" key: {token}'
             )
 
-    def get_token(self, key: str, on_missing_token=None, *on_missing_token_args) -> str:
+    def get_token(
+        self,
+        key: str,
+        *,
+        early_expiry: float = 30.0,
+        on_missing_token=None,
+        **on_missing_token_kwargs,
+    ) -> str:
         """
         Return the bearer token.
         :param key: key identifier of the token
+        :param early_expiry: As the time between the token extraction from cache and the token reception on server side
+        might not higher than one second, on slow networks, token might be expired when received by the actual server,
+        even if still valid when fetched.
+        This is the number of seconds to substract to the actual token expiry. Token will be considered as
+        expired 30 seconds before real expiry by default.
         :param on_missing_token: function to call when token is expired or missing (returning token and expiry tuple)
-        :param on_missing_token_args: arguments of the function
+        :param on_missing_token_kwargs: arguments of the function (key-value arguments)
         :return: the token
         :raise AuthenticationFailed: in case token cannot be retrieved.
         """
@@ -97,7 +117,7 @@ class TokenMemoryCache:
             self._load_tokens()
             if key in self.tokens:
                 bearer, expiry = self.tokens[key]
-                if is_expired(expiry):
+                if _is_expired(expiry, early_expiry):
                     logger.debug(f'Authentication token with "{key}" key is expired.')
                     del self.tokens[key]
                 else:
@@ -109,13 +129,13 @@ class TokenMemoryCache:
         logger.debug("Token cannot be found in cache.")
         if on_missing_token is not None:
             with self.forbid_concurrent_missing_token_function_call:
-                new_token = on_missing_token(*on_missing_token_args)
+                new_token = on_missing_token(**on_missing_token_kwargs)
                 if len(new_token) == 2:  # Bearer token
                     state, token = new_token
-                    self.add_bearer_token(state, token)
+                    self._add_bearer_token(state, token)
                 else:  # Access Token
                     state, token, expires_in = new_token
-                    self.add_access_token(state, token, expires_in)
+                    self._add_access_token(state, token, expires_in)
                 if key != state:
                     logger.warning(
                         f"Using a token received on another key than expected. Expecting {key} but was {state}."
@@ -134,6 +154,7 @@ class TokenMemoryCache:
         raise AuthenticationFailed()
 
     def clear(self):
+        """Remove tokens from the cache."""
         with self.forbid_concurrent_cache_access:
             logger.debug("Clearing token cache.")
             self.tokens = {}